move the samples in to sinks (as iobased_sinks might be too obscure for users). add folder README add mode comments per damccorm comments.
